run-single: Validate step size
This adds a check that the step size is positive, and I've made a small
fix to the simulation algorithm that makes it handle the last time step
properly so that t never exceeds the end time, even if the step size is
greater than the total simulation time.  This closes issue #14.