Support relative FMU paths on Windows (#98)
This fixes #84, enabling users to pass relative FMU paths to `inspect`
and `run-single` on Windows too.  The problem was that we tried to call
`cosim::path_to_file_uri()` even on relative paths, which doesn't make
sense, since a relative path is *already* a relative URI reference. (On
Windows we also have to replace backslashes with forward slashes for
this to be true, which I've also done here.)